reorganizing solvers and constraints for new augmented lagrangian solver
