+++ include/ilqgames/constraint/time_invariant_equality_constraint.h
@@ -36,14 +36,14 @@
 
 ///////////////////////////////////////////////////////////////////////////////
 //
-// (Time-invariant) linear equality constraint, i.e., a^T x - b = 0.
+// Base class for all time-invariant explicit equality constraints.
 //
 ///////////////////////////////////////////////////////////////////////////////
 
-#ifndef ILQGAMES_CONSTRAINT_EXPLICIT_AFFINE_EQUALITY_CONSTRAINT_H
-#define ILQGAMES_CONSTRAINT_EXPLICIT_AFFINE_EQUALITY_CONSTRAINT_H
+#ifndef ILQGAMES_CONSTRAINT_TIME_INVARIANT_EQUALITY_CONSTRAINT_H
+#define ILQGAMES_CONSTRAINT_TIME_INVARIANT_EQUALITY_CONSTRAINT_H
 
-#include <ilqgames/constraint/explicit/time_invariant_equality_constraint.h>
+#include <ilqgames/constraint/equality_constraint.h>
 #include <ilqgames/utils/types.h>
 
 #include <glog/logging.h>
@@ -52,41 +52,30 @@
 
 namespace ilqgames {
 
-class AffineEqualityConstraint : public TimeInvariantEqualityConstraint {
+class TimeInvariantEqualityConstraint : public EqualityConstraint {
  public:
-  ~AffineEqualityConstraint() {}
-  AffineEqualityConstraint(const VectorXf& a, float b,
-                           const std::string& name = "")
-      : TimeInvariantEqualityConstraint(name), a_(a), b_(b) {}
+  virtual ~TimeInvariantEqualityConstraint() {}
 
   // Check if this constraint is satisfied, and optionally return the constraint
   // value, which equals zero if the constraint is satisfied.
-  bool IsSatisfied(const VectorXf& input, float* level) const {
-    CHECK_EQ(input.size(), a_.size());
-    const float value = a_.transpose() * input + b_;
-
-    if (*level) *level = value;
-    return std::abs(value) < constants::kSmallNumber;
+  virtual bool IsSatisfied(const VectorXf& input, float* level) const = 0;
+  bool IsSatisfied(Time t, const VectorXf& input, float* level) const {
+    return IsSatisfied(input, level);
   }
 
   // Quadraticize the constraint value. Do *not* keep a running sum since we
   // keep separate multipliers for each constraint.
-  void Quadraticize(const VectorXf& input, Eigen::Ref<MatrixXf> hess,
+  virtual void Quadraticize(const VectorXf& input, Eigen::Ref<MatrixXf> hess,
+                            Eigen::Ref<VectorXf> grad) const = 0;
+  void Quadraticize(Time t, const VectorXf& input, Eigen::Ref<MatrixXf> hess,
                     Eigen::Ref<VectorXf> grad) const {
-    CHECK_EQ(input.size(), a_.size());
-    CHECK_EQ(hess.rows(), input.size());
-    CHECK_EQ(hess.cols(), input.size());
-    CHECK_EQ(grad.size(), input.size());
+    return Quadraticize(input, hess, grad);
+  };
 
-    // Do nothing to the Hessian - assume that it's already zero.
-    grad = a_;
-  }
-
- private:
-  // Coefficient vector and nominal value.
-  const VectorXf a_;
-  const float b_;
-};  //\class AffineEqualityConstraint
+ protected:
+  explicit TimeInvariantEqualityConstraint(const std::string& name)
+      : EqualityConstraint(name) {}
+};  // namespace ilqgames
 
 }  // namespace ilqgames
