+++ src/three_player_flat_intersection_example.cpp
@@ -57,10 +57,7 @@
 #include <ilqgames/dynamics/single_player_flat_unicycle_4d.h>
 #include <ilqgames/examples/three_player_flat_intersection_example.h>
 #include <ilqgames/geometry/polyline2.h>
-#include <ilqgames/solver/ilq_flat_solver.h>
-#include <ilqgames/solver/lq_feedback_solver.h>
 #include <ilqgames/solver/problem.h>
-#include <ilqgames/solver/solver_params.h>
 #include <ilqgames/utils/initialize_along_route.h>
 #include <ilqgames/utils/solver_log.h>
 #include <ilqgames/utils/strategy.h>
